PoC: expose description to diagnostic handlers
This is a bit of a hack: a JCDiagnostic stores the
message bundle key and arguments, ErrorProne uses
fixed keys and a formatted message. This moves the
formatting inside an object's toString such that a
Log.DiagnosticHandler could detect an ErrorProne
diagnostic and extract the object from the args,
therefore gaining access to the bug details rather
than only the formatted message.